Cleaned up duty cycle calculation

+++ pwm_funcs.h
@@ -9,7 +9,7 @@
 #define	PWM_FUNCS_H
 
 void setup_pwm(uint8_t channel);
-void set_pwm_duty_cycle(uint8_t channel, uint8_t percent);
+void set_pwm_duty_cycle(uint8_t channel, uint32_t percent);
 void pwm_on(uint8_t channel);
 void pwm_off(uint8_t channel);
